+++ PM_QLTBHTD.Application/Services/KhuVucService.cs
@@ -1,4 +1,8 @@
+using Microsoft.EntityFrameworkCore;
 using PM_QLTBHTD.Application.DTOs;
+using PM_QLTBHTD.Application.Exceptions;
+using PM_QLTBHTD.Application.Interfaces;
+using PM_QLTBHTD.Application.Services.IService;
 using PM_QLTBHTD.Domain.Entities;
 using PM_QLTBHTD.Domain.IRepository;
 
@@ -7,10 +11,12 @@ namespace PM_QLTBHTD.Application.Services
     public class KhuVucService : IKhuVucService
     {
         private readonly IKhuVucRepository _repository;
+        private readonly IAppDbContext _db;
 
-        public KhuVucService(IKhuVucRepository repository)
+        public KhuVucService(IKhuVucRepository repository, IAppDbContext db)
         {
             _repository = repository;
+            _db = db;
         }
 
         public async Task<PagedResult<KhuVucDto>> GetPagedAsync(string? search, int page, int? pageSize)
@@ -68,6 +74,10 @@ namespace PM_QLTBHTD.Application.Services
         {
             var entity = await _repository.GetByIdAsync(id);
             if (entity == null) return false;
+
+            var soTramDangDung = await _db.TramDiens.CountAsync(x => x.IDKhuVuc == id);
+            if (soTramDangDung > 0)
+                throw new KhuVucDangSuDungException(id, soTramDangDung);
 
             _repository.Delete(entity);
             await _repository.SaveChangesAsync();
