+++ PM_QLTBHTD.Application/Services/TramDienService.cs
@@ -1,6 +1,8 @@
 using Microsoft.EntityFrameworkCore;
 using PM_QLTBHTD.Application.DTOs;
+using PM_QLTBHTD.Application.Exceptions;
 using PM_QLTBHTD.Application.Interfaces;
+using PM_QLTBHTD.Application.Services.IService;
 using PM_QLTBHTD.Domain.Entities;
 using PM_QLTBHTD.Domain.IRepository;
 
@@ -121,6 +123,10 @@ namespace PM_QLTBHTD.Application.Services
         {
             var entity = await _repository.GetByIdAsync(id);
             if (entity == null) return false;
+
+            var soThietBiDangDung = await _db.ThietBis.CountAsync(x => x.ID_Tram == id);
+            if (soThietBiDangDung > 0)
+                throw new TramDienDangSuDungException(id, soThietBiDangDung);
 
             _repository.Delete(entity);
             await _repository.SaveChangesAsync();
